consumer_5min/requirements.txt
psycopg2-binary==2.9.7
kafka-python==2.0.2